Send SIGTERM by default

+++ src/kill/kill.go
@@ -4,11 +4,18 @@ import (
 	"flag"
 	"os"
 	"strconv"
+	"syscall"
 )
 
 func main() {
+	var killProc bool
+	flag.BoolVar(&killProc, "kill", false, "Send a SIGKILL instead of a SIGTERM.")
 	flag.Parse()
 	pid, _ := strconv.Atoi(flag.Arg(0))
 	proccess, _ := os.FindProcess(pid)
-	proccess.Kill()
+	if killProc {
+		proccess.Kill()
+	} else {
+		proccess.Signal(syscall.SIGTERM)
+	}
 }
